+++ packages/chartjs-plugin/.eslintrc.cjs
@@ -1,0 +1,9 @@
+/* eslint-env node */
+
+module.exports = {
+  root: false, // will be merged with our global config
+  rules: {
+    "no-console": "error",
+    "no-debugger": "error",
+  },
+};