+++ packages/chartjs-plugin/src/stories/BarChart/BarChart.stories.ts
@@ -1,0 +1,12 @@
+import type { Meta, StoryObj } from "@storybook/vue3";
+import BarChart from "./BarChart.vue";
+
+const meta: Meta<typeof BarChart> = {
+  title: "examples/BarChart",
+  component: BarChart,
+};
+
+export default meta;
+type Story = StoryObj<typeof BarChart>;
+
+export const Default = { args: {} } satisfies Story;