+++ packages/chartjs-plugin/src/stories/DoughnutChart/DoughnutChart.ct.tsx
@@ -1,0 +1,8 @@
+import { test } from "@playwright/experimental-ct-vue";
+import { executeChartScreenshotTest } from "../playwright-utils";
+import DoughnutChart from "./DoughnutChart.vue";
+
+test("should render doughnut chart", async ({ mount, page }) => {
+  const component = await mount(DoughnutChart);
+  await executeChartScreenshotTest(page, component, "doughnut");
+});