+++ packages/chartjs-plugin/src/stories/LineChart/LineChart.vue
@@ -1,0 +1,33 @@
+<script lang="ts" setup>
+import type { ChartData } from "chart.js";
+import { Line } from "vue-chartjs";
+import { chartOptions } from "../data";
+
+const labels = ["01.01.2024", "01.02.2024", "01.03.2024", "01.04.2024", "01.05.2024"];
+
+const chartData: ChartData<"line"> = {
+  labels,
+  datasets: [
+    {
+      label: "Dataset A",
+      fill: true,
+      data: [1, 8, 4, 6, 2],
+    },
+    {
+      label: "Dataset B",
+      fill: true,
+      data: [4, 2, 3, 1, 7],
+    },
+  ],
+};
+</script>
+
+<template>
+  <Line :data="chartData" :options="chartOptions" class="chart" />
+</template>
+
+<style lang="scss" scoped>
+.chart {
+  height: 32rem;
+}
+</style>