+++ packages/sit-onyx/src/components/OnyxIconButton/types.ts
@@ -1,5 +1,4 @@
 import type { ButtonType, ButtonVariation } from "../OnyxButton/types";
-import type { IconSize } from "../OnyxIcon/types";
 
 export type OnyxIconButtonProps = {
   /**
@@ -27,9 +26,4 @@ export type OnyxIconButtonProps = {
    * The icon which will be displayed. The custom content in the `default` won't have an effect if the `icon` property is set.
    */
   icon?: string;
-  /**
-   * Icon size. Pixel values will be translated to the according `rem` value by the base of `16px`=`1rem`.
-   * @default 24px
-   */
-  size?: IconSize;
 };