+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.ct.tsx
@@ -82,7 +82,7 @@ test("should show optional marker", async ({ mount }) => {
 });
 
 const STATES = {
-  variant: ["default", "placeholder", "initialValue", "loading"],
+  variant: ["default", "placeholder", "initialValue", "loading", "autofill"],
   writeMode: ["write", "readonly", "disabled"],
   focusState: ["", "hover", "focus"],
 } as const;
@@ -114,6 +114,22 @@ test("should show counter", async ({ mount }) => {
   await expect(component).toHaveScreenshot("counter.png");
 });
 
+test("should have aria-label if label is hidden", async ({ mount, makeAxeBuilder }) => {
+  // ARRANGE
+  const component = await mount(<OnyxInput label="Test label" style="width: 12rem;" hideLabel />);
+
+  // ACT
+  const accessibilityScanResults = await makeAxeBuilder().analyze();
+
+  // ASSERT
+  expect(accessibilityScanResults.violations).toEqual([]);
+
+  // ASSERT
+  await expect(component).not.toContainText("Test label");
+  await expect(component.getByLabel("Test label")).toBeAttached();
+  await expect(component).toHaveScreenshot();
+});
+
 test(
   "State screenshot testing",
   createScreenshotsForAllStates(
@@ -128,11 +144,16 @@ test(
           readonly={writeMode === "readonly"}
           disabled={writeMode === "disabled"}
           loading={variant === "loading"}
+          autocomplete={variant === "autofill" ? "name" : undefined}
           style="width: 12rem;"
         />,
       );
 
       const input = component.getByLabel("Label");
+
+      if (variant == "autofill") {
+        await input.evaluate((node) => node.setAttribute("data-test-autofill", ""));
+      }
 
       if (focusState === "hover") await input.hover();
       if (focusState === "focus") await input.focus();