+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # alpha-test-app
 
+## 0.0.1-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [fd86fa1]
+  - sit-onyx@1.0.0-alpha.46
+
+## 0.0.1-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [27c2bb8]
+  - sit-onyx@1.0.0-alpha.45
+
 ## 0.0.1-alpha.41
 
 ### Patch Changes