+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [fd86fa1]
+  - sit-onyx@1.0.0-alpha.46
+
+## 1.0.0-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [27c2bb8]
+  - sit-onyx@1.0.0-alpha.45
+
 ## 1.0.0-alpha.46
 
 ### Patch Changes